# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Dunda club operating system API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /branches/{branchId}/floor:
    get:
      operationId: getBranchFloor
      tags: [branches]
      summary: Get branch floor and table status
      parameters:
        - $ref: "#/components/parameters/BranchId"
      responses:
        "200":
          description: Floor sections and tables
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Floor"
...
  /floors:
    get:
      operationId: getFloors
      tags: [floors]
      summary: List floors and sections for a branch
      parameters:
        - in: query
          name: branchId
          schema: { type: string }
      responses:
        "200":
          description: Floors with sections
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FloorLayout" }
    post:
      operationId: createFloor
      tags: [floors]
      summary: Create a floor
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FloorInput" }
      responses:
        "201":
          description: Created floor
  /floors/{floorId}:
    patch:
      operationId: updateFloor
      tags: [floors]
      summary: Rename or reorder a floor
      parameters:
        - in: path
          name: floorId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FloorInput" }
      responses:
        "200":
          description: Updated floor
    delete:
      operationId: deleteFloor
      tags: [floors]
      summary: Delete an empty floor
      parameters:
        - in: path
          name: floorId
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Removed
  /floors/{floorId}/sections:
    post:
      operationId: createFloorSection
      tags: [floors]
      summary: Create a section
      parameters:
        - in: path
          name: floorId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FloorSectionInput" }
      responses:
        "201":
          description: Created section
  /floors/sections/{sectionId}:
    patch:
      operationId: updateFloorSection
      tags: [floors]
      summary: Rename a section
      parameters:
        - in: path
          name: sectionId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FloorSectionInput" }
      responses:
        "200":
          description: Updated section
    delete:
      operationId: deleteFloorSection
      tags: [floors]
      summary: Delete a section and its tables
      parameters:
        - in: path
          name: sectionId
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Removed
  /floors/tables:
    get:
      operationId: getFloorTables
      tags: [floors]
      summary: List tables for a branch
      parameters:
        - in: query
          name: branchId
          schema: { type: string }
      responses:
        "200":
          description: Tables
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Table" }
    post:
      operationId: createTable
      tags: [floors]
      summary: Add a table
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TableInput" }
      responses:
        "201":
          description: Created table
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Table" }
  /floors/tables/{tableId}:
    patch:
      operationId: updateTable
      tags: [floors]
      summary: Move, resize, rename or restatus a table
      parameters:
        - in: path
          name: tableId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TableInput" }
      responses:
        "200":
          description: Updated table
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Table" }
    delete:
      operationId: deleteTable
      tags: [floors]
      summary: Delete an unoccupied table
      parameters:
        - in: path
          name: tableId
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Removed
...
components:
  parameters:
    BranchId:
      in: path
      name: branchId
      required: true
      schema:
        type: string
...
  schemas:
...
    Table:
      type: object
      required: [id, name, section, seats, status, total, tabId]
      properties:
        id: { type: string }
        name: { type: string }
        section: { type: string }
        seats: { type: integer }
        status: { type: string, enum: [AVAILABLE, OCCUPIED, RESERVED, PAYMENT_PENDING, CLEANING] }
        total: { type: number }
        tabId: { type: ["string", "null"] }
        customer: { type: ["string", "null"] }
        x: { type: integer }
        y: { type: integer }
        width: { type: integer }
        height: { type: integer }
    FloorSection:
      type: object
      required: [id, name, tables]
      properties:
        id: { type: string }
        name: { type: string }
        tables:
          type: array
          items: { $ref: "#/components/schemas/Table" }
    Floor:
      type: object
      required: [branchId, sections]
      properties:
        branchId: { type: string }
        sections:
          type: array
          items: { $ref: "#/components/schemas/FloorSection" }
...
    FloorLayout:
      type: object
      required: [branchId, floors]
      properties:
        branchId: { type: string }
        floors:
          type: array
          items:
            type: object
            required: [id, name, sortOrder, sections]
            properties:
              id: { type: string }
              name: { type: string }
              sortOrder: { type: integer }
              sections:
                type: array
                items:
                  type: object
                  required: [id, name, color, sortOrder]
                  properties:
                    id: { type: string }
                    name: { type: string }
                    color: { type: string }
                    sortOrder: { type: integer }
    FloorInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        sortOrder: { type: integer }
        branchId: { type: string }
    FloorSectionInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        color: { type: string }
        sortOrder: { type: integer }
    TableInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        section: { type: string, minLength: 1 }
        floorSectionId: { type: ["string", "null"] }
        seats: { type: integer, minimum: 1 }
        x: { type: integer }
        y: { type: integer }
        width: { type: integer, minimum: 40 }
        height: { type: integer, minimum: 40 }
        status:
          type: string
          enum: [AVAILABLE, OCCUPIED, RESERVED, PAYMENT_PENDING, CLEANING]
        branchId: { type: string }
        floorId: { type: string }
...